requirements.txt
# Web Framework
Flask==3.0.0
Flask-Cors==4.0.0

# Image Processing
Pillow==10.0.0

# PDF Processing
pypdf==5.8.0

# OCR and Text Processing
openai==1.97.0
ftfy==6.3.1

# Environment Variables
python-dotenv==1.0.0

# Additional Libraries
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
google-generativeai==0.3.2

# Note: Poppler is a system requirement, not a Python package
# For Windows: Download from https://github.com/oschwartz10612/poppler-windows/releases/
# For macOS: brew install poppler
# For Linux: sudo apt-get install poppler-utils
